requirements.txt
appdirs==1.4.3
packaging==16.8
pyldap==2.4.28
pyparsing==2.2.0
PyYAML==3.12
six==1.10.0
